fix: Merge all system messages
Fixes #1860
